Rename options.py -> pipeline_options.py
