+++ sdks/python/apache_beam/examples/cookbook/bigquery_side_input.py
@@ -35,8 +35,8 @@ import apache_beam as beam
 
 from apache_beam.pvalue import AsList
 from apache_beam.pvalue import AsSingleton
-from apache_beam.utils.options import PipelineOptions
-from apache_beam.utils.options import SetupOptions
+from apache_beam.utils.pipeline_options import PipelineOptions
+from apache_beam.utils.pipeline_options import SetupOptions
 
 
 def create_groups(group_ids, corpus, word, ignore_corpus, ignore_word):
