+++ sdks/python/apache_beam/examples/cookbook/coders.py
@@ -35,8 +35,8 @@ import json
 import logging
 
 import apache_beam as beam
-from apache_beam.utils.options import PipelineOptions
-from apache_beam.utils.options import SetupOptions
+from apache_beam.utils.pipeline_options import PipelineOptions
+from apache_beam.utils.pipeline_options import SetupOptions
 
 
 class JsonCoder(object):
