+++ sdks/python/apache_beam/examples/cookbook/custom_ptransform.py
@@ -27,7 +27,7 @@ import logging
 
 import apache_beam as beam
 
-from apache_beam.utils.options import PipelineOptions
+from apache_beam.utils.pipeline_options import PipelineOptions
 
 
 # pylint doesn't understand our pipeline syntax:
