+++ sdks/python/apache_beam/examples/cookbook/mergecontacts.py
@@ -36,8 +36,8 @@ import logging
 import re
 
 import apache_beam as beam
-from apache_beam.utils.options import PipelineOptions
-from apache_beam.utils.options import SetupOptions
+from apache_beam.utils.pipeline_options import PipelineOptions
+from apache_beam.utils.pipeline_options import SetupOptions
 
 
 def run(argv=None, assert_results=None):
