+++ sdks/python/apache_beam/examples/cookbook/multiple_output_pardo.py
@@ -54,8 +54,8 @@ import re
 
 import apache_beam as beam
 from apache_beam import pvalue
-from apache_beam.utils.options import PipelineOptions
-from apache_beam.utils.options import SetupOptions
+from apache_beam.utils.pipeline_options import PipelineOptions
+from apache_beam.utils.pipeline_options import SetupOptions
 
 
 class SplitLinesToWordsFn(beam.DoFn):
