+++ sdks/python/apache_beam/examples/wordcount.py
@@ -24,8 +24,8 @@ import logging
 import re
 
 import apache_beam as beam
-from apache_beam.utils.options import PipelineOptions
-from apache_beam.utils.options import SetupOptions
+from apache_beam.utils.pipeline_options import PipelineOptions
+from apache_beam.utils.pipeline_options import SetupOptions
 
 
 empty_line_aggregator = beam.Aggregator('emptyLines')
