+++ sdks/python/apache_beam/internal/apiclient_test.py
@@ -18,7 +18,7 @@
 
 import unittest
 
-from apache_beam.utils.options import PipelineOptions
+from apache_beam.utils.pipeline_options import PipelineOptions
 from apache_beam.runners.dataflow_runner import DataflowRunner
 from apache_beam.internal import apiclient
