+++ sdks/python/apache_beam/io/bigquery.py
@@ -120,7 +120,7 @@ from apache_beam.internal.json_value import to_json_value
 from apache_beam.runners.dataflow.native_io import iobase as dataflow_io
 from apache_beam.transforms.display import DisplayDataItem
 from apache_beam.utils import retry
-from apache_beam.utils.options import GoogleCloudOptions
+from apache_beam.utils.pipeline_options import GoogleCloudOptions
 
 # Protect against environments where bigquery library is not available.
 # pylint: disable=wrong-import-order, wrong-import-position
