+++ sdks/python/apache_beam/typehints/typed_pipeline_test.py
@@ -25,8 +25,8 @@ from apache_beam import pvalue
 from apache_beam import typehints
 from apache_beam.transforms.util import assert_that, equal_to
 from apache_beam.typehints import WithTypeHints
-from apache_beam.utils.options import OptionsContext
-from apache_beam.utils.options import PipelineOptions
+from apache_beam.utils.pipeline_options import OptionsContext
+from apache_beam.utils.pipeline_options import PipelineOptions
 
 # These test often construct a pipeline as value | PTransform to test side
 # effects (e.g. errors).
