+++ sdks/python/apache_beam/utils/pipeline_options_validator.py
@@ -20,13 +20,13 @@
 import re
 
 from apache_beam.internal import pickler
-from apache_beam.utils.options import DebugOptions
-from apache_beam.utils.options import GoogleCloudOptions
-from apache_beam.utils.options import SetupOptions
-from apache_beam.utils.options import StandardOptions
-from apache_beam.utils.options import TestOptions
-from apache_beam.utils.options import TypeOptions
-from apache_beam.utils.options import WorkerOptions
+from apache_beam.utils.pipeline_options import DebugOptions
+from apache_beam.utils.pipeline_options import GoogleCloudOptions
+from apache_beam.utils.pipeline_options import SetupOptions
+from apache_beam.utils.pipeline_options import StandardOptions
+from apache_beam.utils.pipeline_options import TestOptions
+from apache_beam.utils.pipeline_options import TypeOptions
+from apache_beam.utils.pipeline_options import WorkerOptions
 
 
 class PipelineOptionsValidator(object):
