+++ sdks/python/apache_beam/utils/pipeline_options_validator_test.py
@@ -21,7 +21,7 @@ import logging
 import unittest
 
 from apache_beam.internal import pickler
-from apache_beam.utils.options import PipelineOptions
+from apache_beam.utils.pipeline_options import PipelineOptions
 from apache_beam.utils.pipeline_options_validator import PipelineOptionsValidator
 from hamcrest.core.base_matcher import BaseMatcher
